+++ coroutines/src/commonTest/kotlin/SpecialMutableStateFlowTests.kt
@@ -1,4 +1,4 @@
-import dev.inmo.micro_utils.coroutines.SpecialMutableStateFlow
+import dev.inmo.micro_utils.coroutines.MutableRedeliverStateFlow
 import dev.inmo.micro_utils.coroutines.asDeferred
 import dev.inmo.micro_utils.coroutines.subscribe
 import kotlinx.coroutines.Job
@@ -8,17 +8,17 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
-class SpecialMutableStateFlowTests {
+class MutableRedeliverStateFlowTests {
     @Test
     fun simpleTest() = runTest {
-        val specialMutableStateFlow = SpecialMutableStateFlow(0)
+        val specialMutableStateFlow = MutableRedeliverStateFlow(0)
         specialMutableStateFlow.value = 1
         specialMutableStateFlow.first { it == 1 }
         assertEquals(1, specialMutableStateFlow.value)
     }
     @Test
     fun specialTest() = runTest {
-        val specialMutableStateFlow = SpecialMutableStateFlow(0)
+        val specialMutableStateFlow = MutableRedeliverStateFlow(0)
         lateinit var subscriberJob: Job
         subscriberJob = specialMutableStateFlow.subscribe(this) {
             when (it) {